Extract Value to its own file
